implement refresh_tokens

+++ lib/OAuth/AuthorizationServer.php
@@ -13,6 +13,9 @@ interface IOAuthStorage {
     public function getAuthorizationCode   ($authorizationCode, $redirectUri);
     public function deleteAuthorizationCode($authorizationCode, $redirectUri);
 
+    public function getRefreshToken        ($refreshToken);
+    public function storeRefreshToken      ($refreshToken, $clientId, $resourceOwnerId, $scope);
+
     public function getResourceOwner       ($resourceOwnerId);
     public function storeResourceOwner     ($resourceOwnerId, $resourceOwnerDisplayName);
 
@@ -160,7 +163,6 @@ class AuthorizationServer {
                                          "user_agent_based_application" => array ("token"));
 
         if(!in_array($responseType, $allowedClientProfiles[$client->type])) {
-
             throw new ClientException("unsupported_response_type", "response_type not supported by client profile", $client, $state);
         }
 
@@ -260,25 +262,39 @@ class AuthorizationServer {
 
     public function token(array $post, $authorizationHeader) {
         // exchange authorization code for access token
-        $grantType   = self::getParameter($post, 'grant_type');
-        $code        = self::getParameter($post, 'code');
-        $redirectUri = self::getParameter($post, 'redirect_uri');
+        $grantType    = self::getParameter($post, 'grant_type');
+        $code         = self::getParameter($post, 'code');
+        $redirectUri  = self::getParameter($post, 'redirect_uri');
+        $refreshToken = self::getParameter($post, 'refresh_token');
 
         if(NULL === $grantType) {
             throw new TokenException("invalid_request: the grant_type parameter is missing");
         }
-        if("authorization_code" !== $grantType) {
+        if("authorization_code" !== $grantType && "refresh_token" !== $grantType) {
             throw new TokenException("unsupported_grant_type: the requested grant type is not supported");
         }
-        if(NULL === $code) {
-            throw new TokenException("invalid_request: the code parameter is missing");
-        }
-        $result = $this->_storage->getAuthorizationCode($code, $redirectUri);
-        if(FALSE === $result) {
-            throw new TokenException("invalid_grant: the authorization code was not found");
-        }
-        if(time() > $result->issue_time + 600) {
-            throw new TokenException("invalid_grant: the authorization code expired");
+
+        if("authorization_code" === $grantType) {
+            // authorization_code
+            if(NULL === $code) {
+                throw new TokenException("invalid_request: the code parameter is missing");
+            }
+            $result = $this->_storage->getAuthorizationCode($code, $redirectUri);
+            if(FALSE === $result) {
+                throw new TokenException("invalid_grant: the authorization code was not found");
+            }
+            if(time() > $result->issue_time + 600) {
+                throw new TokenException("invalid_grant: the authorization code expired");
+            }
+        } else {
+            // refresh_token
+            if(NULL === $refreshToken) {
+                throw new TokenException("invalid_request: the refresh_token parameter is missing");
+            }
+            $result = $this->_storage->getRefreshToken($refreshToken);        
+            if(FALSE === $result) {
+                throw new TokenException("invalid_grant: the refresh_token was not found");
+            }
         }
 
         $client = $this->_storage->getClient($result->client_id);
@@ -302,13 +318,34 @@ class AuthorizationServer {
                 }
             }
         }
-        // we need to be able to delete, otherwise someone else was first!
-        if(FALSE === $this->_storage->deleteAuthorizationCode($code, $redirectUri)) {
-            throw new TokenException("invalid_grant: this grant was already used");
+
+        if($client->id !== $result->client_id) {
+            throw new TokenException("invalid_grant: grant was not issued to this client");
         }
-        $token = $this->_storage->getAccessToken($result->access_token);
+
+        if("authorization_code" === $grantType) {
+            // we need to be able to delete, otherwise someone else was first!
+            if(FALSE === $this->_storage->deleteAuthorizationCode($code, $redirectUri)) {
+                throw new TokenException("invalid_grant: this grant was already used");
+            }
+
+            // FIXME: maybe we should only create the access token here, so we don't need to 
+            // add it to the AuthorizationCode table... Or was there a special reason to do it there?
+            $token = $this->_storage->getAccessToken($result->access_token);
+
+            // create a new refresh token
+            $token->refresh_token = self::randomHex(16);
+            $this->_storage->storeRefreshToken($token->refresh_token, $token->client_id, $token->resource_owner_id, $token->scope);
+        } else {
+            // refresh_token
+
+            // create a new access token
+            $accessToken = self::randomHex(16);
+            $this->_storage->storeAccessToken($accessToken, time(), $result->client_id, $result->resource_owner_id, $result->scope, $this->_c->getValue('accessTokenExpiry'));
+            $token = $this->_storage->getAccessToken($accessToken);
+        }
+
         $token->expires_in = $token->issue_time + $token->expires_in - time();
-        // FIXME: move token_type to DB
         $token->token_type = 'bearer';
         // filter unwanted response parameters
         $responseParameters = array("access_token", "token_type", "expires_in", "refresh_token", "scope");
